Smart_Blind_Stick/smart_cane_server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4